메인페이지 UI이슈 작업
