+++ src/components/DevToolsList.jsx
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import RESP from "mockAPI/reponse";
 import DevTool from "./DevTool";
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
