+++ src/mockAPI/response.js
@@ -1,0 +1,36 @@
+const RESP = {
+  LoginUserName: "로그인된 유저이름",
+
+  data: {
+    responseArticles: [
+      {
+        id: 1,
+        title: "제목1",
+        content: "내용1",
+        username: "admin1",
+        category: "하드웨어",
+        createAt: "2022년 8월 12일 3시 33분",
+        isMyArticles: true,
+      },
+      {
+        id: 2,
+        title: "제목2",
+        content: "내용2",
+        username: "admin2",
+        category: "소프트웨어",
+        createAt: "2022년 8월 12일 3시 35분",
+        isMyArticles: true,
+      },
+      {
+        id: 3,
+        title: "제목3",
+        content: "내용3",
+        username: "admin3",
+        category: "하드웨어",
+        createAt: "2022년 8월 12일 3시 55분",
+        isMyArticles: true,
+      },
+    ],
+  },
+};
+export default RESP;
